bot/requirements.txt
aiogram>=3.4,<4
httpx>=0.27
APScheduler>=3.10
aiosqlite>=0.20
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
GitPython>=3.1
aiohttp>=3.9
